ajout vehicle api key

+++ src/v1/vehicle.php
@@ -4,6 +4,10 @@ use \Psr\Http\Message\ServerRequestInterface as Request;
 $app->group('/vehicle', function () use ($app){
     $app->post('/add', function (Request $request, Response $response, $args) {
         $data = $request->getParsedBody();
+        $api = new Api();
+        if (!$api->checkout($data['api_key'])){
+            return $response->withJson(['success' => false]);
+        }
         $dev = new Vehicles();
         $dev->car_brand = $data['marque'];
         $dev->model = $data['model'];
@@ -19,6 +23,10 @@ $app->group('/vehicle', function () use ($app){
 
     $app->post('/delete', function (Request $request, Response $response, $args) {
         $data = $request->getParsedBody();
+        $api = new Api();
+        if (!$api->checkout($data['api_key'])){
+            return $response->withJson(['success' => false]);
+        }
          Vehicles::where('idvehicle', $data['id'])->delete();
         return $response->withJson(['success' => true, 'data' => "test"]);
     });
@@ -29,6 +37,10 @@ $app->group('/vehicle', function () use ($app){
 
     $app->post('/update', function (Request $request, Response $response, $args) {
         $data = $request->getParsedBody();
+        $api = new Api();
+        if (!$api->checkout($data['api_key'])){
+            return $response->withJson(['success' => false]);
+        }
         foreach ($data['data'] as $key => $item){
             Vehicles::where('idvehicle', $data['id'])
                 ->update([$key => $item]);
